cook-rag/pyproject.toml
[project]
name = "cook-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.6.0",
    "dagit>=1.10.14",
    "dagster>=1.10.14",
    "google-cloud-storage>=3.1.0",
    "huggingface-hub>=0.31.1",
    "openai>=1.78.0",
    "pinecone>=6.0.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "tiktoken>=0.9.0",
]
